+++ CHANGELOG.md
@@ -11,6 +11,31 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Fixes
 
+## v3.3.0 (2024-04-01)
+
+### New
+- support to generate schema for manifests from CLI (`seedfarmer list schema`)
+- added commithash persistence support for modules sourced from git
+  - recorded in module manifest (`commit_hash`)
+  - recorded in module metadata as `SeedFarmerModuleCommitHash` and can be fetched
+
+### Changes
+- renaming the threads spawned for deploy / destroy to indicate the module being worked on
+- adding detailed docs for CLI parameters
+- complete support to delete the seedkit on `seedfarmer destroy` command
+- adding verbose messaging to session manager and hints to reconcile session issues
+- adding info for destroy and list deployments when no deployments found
+- refactored git support logic to separate python file
+- added verbose messaging related to git issues
+- updated `black~=24.3.0` in requirements-dev as per dependabot
+- updated `werkzeug~=2.3.8` in requirements-dev as per dependabot
+- removed explicit support for python 3.7
+  - this was only due to support for black in requirements-dev, but was also removed from setup.py
+- loosened restrictions on `typing-requirements` library 
+
+
+### Fixes
+- Add schema validation step checking that either `value` or `value_from` is present for each parameter
 
 ## v3.2.2 (2024-02-27)